# src/swagger/swagger.yaml
consumes:
- application/json
definitions:
...
  Findbugs:
    description: Description of which Findbugs bugs are checked.
    example:
      name: Nulla ipsa.
    properties:
      name:
        example: Nulla ipsa.
        type: string
    title: Findbugs
    type: object
...
  UserMt:
    description: UserMt media type (default view)
    example:
      email: stewart_denesik@maggiowiza.net
      id: 2.7364505e+18
      name: Modi ut.
      phone_number: Dignissimos in.
    properties:
      email:
        description: User email
        example: stewart_denesik@maggiowiza.net
        format: email
        type: string
      id:
        example: 2.7364505e+18
        format: int64
        type: integer
      name:
        example: Modi ut.
        type: string
      phone_number:
        example: Dignissimos in.
        type: string
    title: 'Mediatype identifier: application/user.mt; view=default'
    type: object
  UserMtGithub:
    description: UserMt media type (github view)
    example:
      callback: Dignissimos aliquid saepe.
      email: stewart_denesik@maggiowiza.net
      id: 2.7364505e+18
      name: Modi ut.
      phone_number: Dignissimos in.
    properties:
      callback:
        example: Dignissimos aliquid saepe.
        type: string
      email:
        description: User email
        example: stewart_denesik@maggiowiza.net
        format: email
        type: string
      id:
        example: 2.7364505e+18
        format: int64
        type: integer
      name:
        example: Modi ut.
        type: string
      phone_number:
        example: Dignissimos in.
        type: string
    title: 'Mediatype identifier: application/user.mt; view=github'
    type: object
  bug_profile:
    description: A bug profile
    example:
      checkstyle:
        AbbreviationAsWordInName: false
        AbstractClassName: false
        AnnotationLocation: true
        AnnotationUseStyle: true
        AnonInnerLength: false
        ArrayTrailingComma: true
        ArrayTypeStyle: true
        AtclauseOrder: true
        AvoidEscapedUnicodeCharacters: true
        AvoidInlineConditionals: true
        AvoidNestedBlocks: true
        AvoidStarImport: false
        AvoidStaticImport: true
        BooleanExpressionComplexity: true
        CatchParameterName: false
        ClassDataAbstractionCoupling: false
        ClassFanOutComplexity: true
        ClassTypeParameterName: false
        CommentsIndentation: false
        ConstantName: true
        CovariantEquals: true
        CustomImportOrder: false
        CyclomaticComplexity: false
        DeclarationOrder: true
        DefaultComesLast: true
        DescendantToken: true
        DesignForExtension: false
        EmptyBlock: false
        EmptyCatchBlock: true
        EmptyForInitializerPad: true
        EmptyForIteratorPad: true
        EmptyLineSeparator: false
        EmptyStatement: false
        EqualsAvoidNull: true
        EqualsHashCode: false
        ExecutableStatementCount: true
        ExplicitInitialization: true
        FallThrough: true
        FileLength: false
        FileTabCharacter: false
        FinalClass: false
        FinalLocalVariable: false
        FinalParameters: true
        GenericWhitespace: false
        Header: false
        HiddenField: true
        HideUtilityClassConstructor: false
        IllegalCatch: true
        IllegalImport: true
        IllegalInstantiation: false
        IllegalThrows: true
        IllegalToken: true
        IllegalTokenText: true
        IllegalType: true
        ImportControl: false
        ImportOrder: true
        Indentation: true
        InnerAssignment: true
        InnerTypeLast: false
        InterfaceIsType: false
        InterfaceTypeParameterName: true
        JavaNCSS: true
        JavadocMethod: false
        JavadocPackage: false
        JavadocParagraph: false
        JavadocStyle: true
        JavadocTagContinuationIndentation: true
        JavadocType: true
        JavadocVariable: true
        LeftCurly: true
        LineLength: false
        LocalFinalVariableName: false
        LocalVariableName: true
        MagicNumber: false
        MemberName: true
        MethodCount: true
        MethodLength: false
        MethodName: false
        MethodParamPad: true
        MethodTypeParameterName: false
        MissingCtor: false
        MissingDeprecated: false
        MissingOverride: false
        MissingSwitchDefault: true
        ModifiedControlVariable: true
        ModifierOrder: false
        MultipleStringLiterals: false
        MultipleVariableDeclarations: false
        MutableException: false
        NPathComplexity: false
        NeedBraces: true
        NestedForDepth: true
        NestedIfDepth: true
        NestedTryDepth: false
        NewlineAtEndOfFile: true
        NoClone: false
        NoFinalizer: true
        NoLineWrap: false
        NoWhitespaceAfter: true
        NoWhitespaceBefore: true
        NonEmptyAtclauseDescription: true
        OneStatementPerLine: true
        OneTopLevelClass: true
        OperatorWrap: false
        OuterTypeFilename: false
        OuterTypeNumber: false
        OverloadMethodsDeclarationOrder: false
        PackageAnnotation: false
        PackageDeclaration: false
        PackageName: true
        ParameterAssignment: true
        ParameterName: true
        ParameterNumber: true
        ParenPad: false
        RedundantImport: true
        RedundantModifier: false
        Regexp: false
        RegexpHeader: true
        RegexpMultiline: false
        RegexpOnFilename: true
        RegexpSingleline: false
        RegexpSinglelineJava: false
        RequireThis: false
        ReturnCount: true
        RightCurly: true
        SeparatorWrap: true
        SimplifyBooleanExpression: true
        SimplifyBooleanReturn: false
        SingleLineJavadoc: false
        SingleSpaceSeparator: true
        StaticVariableName: true
        StringLiteralEquality: true
        SummaryJavadoc: true
        SuperClone: false
        SuperFinalize: false
        SuppressWarnings: false
        SuppressWarningsHolder: false
        ThrowsCount: true
        TodoComment: false
        TrailingComment: false
        Translation: false
        TypeName: true
        TypecastParenPad: true
        UncommentedMain: true
        UniqueProperties: false
        UnnecessaryParentheses: false
        UnusedImports: true
        UpperEll: true
        VariableDeclarationUsageDistance: true
        VisibilityModifier: true
        WhitespaceAfter: true
        WhitespaceAround: false
        WriteTag: true
      findbugs:
        name: Nulla ipsa.
      name: Et aspernatur.
    properties:
      checkstyle:
        $ref: '#/definitions/Checkstyle'
      findbugs:
        $ref: '#/definitions/Findbugs'
      name:
        example: Et aspernatur.
        type: string
    title: bug_profile
    type: object
  user:
    description: payload used to sign up a user
    example:
      email: patsy_prohaska@breitenberg.org
      name: Et nostrum est.
      password: Ut enim ea voluptas magni.
      phone_number: Et minus odio.
    properties:
      email:
        description: User email
        example: patsy_prohaska@breitenberg.org
        format: email
        type: string
      name:
        example: Et nostrum est.
        type: string
      password:
        example: Ut enim ea voluptas magni.
        type: string
      phone_number:
        example: Et minus odio.
        type: string
    title: user
    type: object
host: localhost:8081
info:
  description: The next big thing
  title: Gradebook
  version: ""
paths:
...
  /api/bug-profile:
    get:
      description: Show all of my bug profiles
      operationId: bug_profile#list
      responses:
        "200":
          description: OK
      schemes:
      - http
      summary: list bug_profile
      tags:
      - bug_profile
    post:
      description: Make me a new bug profile
      operationId: bug_profile#create
      parameters:
      - description: A bug profile
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bug_profile'
      responses:
        "200":
          description: OK
      schemes:
      - http
      summary: create bug_profile
      tags:
      - bug_profile
  /api/bug-profile/{profileID}:
    delete:
      description: I don't want this bug profile anymore
      operationId: bug_profile#delete
      parameters:
      - description: The profile's unique identifier
        in: path
        name: profileID
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      schemes:
      - http
      summary: delete bug_profile
      tags:
      - bug_profile
    get:
      description: Show a single bug profile
      operationId: bug_profile#show
      parameters:
      - description: The profile's unique identifier
        in: path
        name: profileID
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      schemes:
      - http
      summary: show bug_profile
      tags:
      - bug_profile
    put:
      description: Update my pre-existing bug profile
      operationId: bug_profile#update
      parameters:
      - description: The profile's unique identifier
        in: path
        name: profileID
        required: true
        type: integer
      - description: A bug profile
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bug_profile'
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      schemes:
      - http
      summary: update bug_profile
      tags:
      - bug_profile
...
  /api/user:
    post:
      description: Sign up for the first time
      operationId: user#create
      parameters:
      - description: payload used to sign up a user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/user'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserMtGithub'
        "500":
          description: Internal Server Error
      schemes:
      - http
      summary: create user
      tags:
      - user
  /api/user/{userID}:
    get:
      description: Get the detials about this particular account
      operationId: user#read
      parameters:
      - description: The user's unique identifier
        in: path
        name: userID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserMt'
        "500":
          description: Internal Server Error
      schemes:
      - http
      summary: read user
      tags:
      - user
    put:
      description: Adjust your account settings
      operationId: user#update
      parameters:
      - description: The user's unique identifier
        in: path
        name: userID
        required: true
        type: integer
      - description: payload used to sign up a user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/user'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserMt'
        "500":
          description: Internal Server Error
      schemes:
      - http
      summary: update user
      tags:
      - user
produces:
- application/json
...
schemes:
- http
securityDefinitions:
  jwt:
    description: |-
      For routes that require an authorization token

      **Token URL**: https://example.com/token

      **Security Scopes**:
        * `profiles:read`: Can read profiles
        * `profiles:write`: Can update profiles
        * `projects:read`: Can read projects
        * `projects:write`: Can update projects
        * `settings:read`: Can read settings
        * `settings:write`: Can update settings
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"